+++ lib/widgets/transaction_list.dart
@@ -1,0 +1,57 @@
+import 'package:flutter/material.dart';
+import '../model/transaction.dart';
+
+import 'package:intl/intl.dart';
+
+class TransactionList extends StatelessWidget {
+  final List<Transaction> transactions;
+
+  const TransactionList({required this.transactions});
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: transactions.map((tx) {
+        return Card(
+          child: Row(
+            children: [
+              Container(
+                margin: EdgeInsets.symmetric(vertical: 10, horizontal: 15),
+                decoration: BoxDecoration(
+                  border: Border.all(color: Colors.purple, width: 2)
+                ),
+                padding: EdgeInsets.all(10),
+                child: Text(
+                  '¥: ${tx.amount.toString()}',
+                  style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 20,
+                    color: Colors.purple
+                  ),
+                ),
+              ),
+              Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    tx.title,
+                    style: TextStyle(
+                      color: Colors.black,
+                      fontWeight: FontWeight.bold,
+                      fontSize: 16,
+                    ),
+                  ),
+                  Text(
+                    DateFormat('yyyy/MM/dd').format(tx.date),
+                    style: TextStyle(
+                      color: Colors.grey
+                    ),
+                  ),
+                ],
+              )
+            ],
+          ),
+        );
+      }).toList(),
+    );
+  }
+}
